move handler tests which are actually testing everything to server package
